requirements.txt
# Core Data Science
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
scipy>=1.12.0
joblib>=1.3.0

# Deep Learning
torch>=2.2.0
torch-geometric>=2.4.0

# FastAPI & Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0

# API Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate Limiting & Monitoring
slowapi>=0.1.8
prometheus-client>=0.18.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Logging
python-json-logger>=2.0.0

# Explainability
shap>=0.44.0
lime>=0.2.0

# Frontend & Visualization
streamlit>=1.31.0
plotly>=5.18.0
networkx>=3.2.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0